Variable "data" in method "u_byte_to_float" is type of int instead short

+++ src/smart_road_radar_utils.hpp
@@ -216,7 +216,8 @@ u_byte_t calculate_checksum(frame target_frame) {
  * \return число типа float
  */
 float u_byte_to_float(const u_byte_t u_bytes[2]) {
-    short data = (u_bytes[1] << 8) | u_bytes[0];
+    // TODO: Проверить это место на реальном радаре
+    int data = (u_bytes[1] << 8) | u_bytes[0];
     return (float) data * SCALE;
 }
